feat: new model path

+++ pdf2zh/scanned/ infrastructure/custom_surya_recognition/__init__.py
@@ -43,7 +43,7 @@ from custom_surya_recognition.cache import (
     ContinuousBatchingQuantizedCache,
 )
 from custom_surya_recognition.settings import settings
-from surya.logging import get_logger, configure_logging
+from custom_surya_recognition.logging import get_logger, configure_logging
 
 configure_logging()
 logger = get_logger()
@@ -81,7 +81,7 @@ class RecognitionPredictor(BasePredictor):
     default_batch_sizes = {"cpu": 32, "mps": 64, "cuda": 256, "xla": 128}
     encoder_chunk_size: int = 4096  # Default chunk size
     encoder_chunk_sizes = {"cpu": 4096, "mps": 4096, "cuda": 32768, "xla": 32768}
-    min_prefill_ratio: int = 0.2
+    min_prefill_ratio: float = 0.2
     min_trim_length: int = 50
     tasks = {
         TaskNames.ocr_with_boxes: {
